+++ Allfiles/Labs/14-Run an IoT Edge device in restricted network and offline/Starter/ChildIoTDevice/CaveDevice.csproj
@@ -9,8 +9,4 @@
     <PackageReference Include="Microsoft.Azure.Devices.Client" Version="1.*" />
   </ItemGroup>
 
-  <ItemGroup>
-    <None Update="azure-iot-test-only.root.ca.cert.pem" CopyToOutputDirectory="PreserveNewest" />
-  </ItemGroup>
-
 </Project>
