--- Allfiles/Labs/14-Run an IoT Edge device in restricted network and offline/Starter/ChildIoTDevice/ChildIoTDevice.cs
@@ -1,104 +1,0 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using Newtonsoft.Json;
-using System;
-using System.Globalization;
-using System.IO;
-using System.Security.Cryptography.X509Certificates;
-using System.Text;
-using System.Threading.Tasks;
-using Microsoft.Azure.Devices.Client;
-
-namespace ChildIoTDevice
-{
-    class SimulatedDevice
-    {
-        private static DeviceClient s_deviceClient;
-
-        // The device connection string to authenticate the device with your IoT Edge Gateway.
-        private readonly static string s_connectionString = "{Your device connection string here}";
-        
-        // Async method to send simulated telemetry
-        private static async void SendDeviceToCloudMessagesAsync()
-        {
-            // Initial telemetry values
-            double minTemperature = 20;
-            double minHumidity = 60;
-            Random rand = new Random();
-
-            while (true)
-            {
-                double currentTemperature = minTemperature + rand.NextDouble() * 15;
-                double currentHumidity = minHumidity + rand.NextDouble() * 20;
-
-                // Create JSON message
-                var telemetryDataPoint = new
-                {
-                    temperature = currentTemperature,
-                    humidity = currentHumidity
-                };
-                var messageString = JsonConvert.SerializeObject(telemetryDataPoint);
-                var message = new Message(Encoding.ASCII.GetBytes(messageString));
-
-                // Add a custom application property to the message.
-                // An IoT hub can filter on these properties without access to the message body.
-                message.Properties.Add("temperatureAlert", (currentTemperature > 30) ? "true" : "false");
-
-                // Send the telemetry message
-                await s_deviceClient.SendEventAsync(message);
-
-                Console.WriteLine("{0} > Sending message: {1}", DateTime.Now, messageString);
-
-                await Task.Delay(1000);
-            }
-        }
-        private static void Main(string[] args)
-        {
-            Console.WriteLine("IoT Hub Quickstarts #1 - Simulated device. Ctrl-C to exit.\n");
-
-            InstallCACert();
-
-            // Connect to the IoT hub using the MQTT protocol
-            s_deviceClient = DeviceClient.CreateFromConnectionString(s_connectionString, TransportType.Mqtt);
-            SendDeviceToCloudMessagesAsync();
-            Console.ReadLine();
-        }
-
-        /// <summary>
-        /// Add certificate in local cert store for use by downstream device
-        /// client for secure connection to IoT Edge runtime.
-        ///
-        ///    Note: On Windows machines, if you have not run this from an Administrator prompt,
-        ///    a prompt will likely come up to confirm the installation of the certificate.
-        ///    This usually happens the first time a certificate will be installed.
-        /// </summary>
-        static void InstallCACert()
-        {
-            string trustedCACertPath = "azure-iot-test-only.root.ca.cert.pem";
-            if (!string.IsNullOrWhiteSpace(trustedCACertPath))
-            {
-                Console.WriteLine("User configured CA certificate path: {0}", trustedCACertPath);
-                if (!File.Exists(trustedCACertPath))
-                {
-                    // cannot proceed further without a proper cert file
-                    Console.WriteLine("Certificate file not found: {0}", trustedCACertPath);
-                    throw new InvalidOperationException("Invalid certificate file.");
-                }
-                else
-                {
-                    Console.WriteLine("Attempting to install CA certificate: {0}", trustedCACertPath);
-                    X509Store store = new X509Store(StoreName.Root, StoreLocation.CurrentUser);
-                    store.Open(OpenFlags.ReadWrite);
-                    store.Add(new X509Certificate2(X509Certificate.CreateFromCertFile(trustedCACertPath)));
-                    Console.WriteLine("Successfully added certificate: {0}", trustedCACertPath);
-                    store.Close();
-                }
-            }
-            else
-            {
-                Console.WriteLine("trustedCACertPath was not set or null, not installing any CA certificate");
-            }
-        }
-    }
-}
